[NFC] Move CPU specific utils to LLVMCPU/
