+++ compiler/src/iree/compiler/Codegen/Utils/Utils.h
@@ -62,39 +62,9 @@ Optional<BoolAttr> getConfigBoolAttr(IREE::HAL::ExecutableTargetAttr targetAttr,
 Optional<llvm::Triple> getTargetTriple(
     IREE::HAL::ExecutableTargetAttr targetAttr);
 
-/// Returns the CPU target features associated with the `targetAttr`, if set.
-Optional<StringRef> getCpuFeatures(IREE::HAL::ExecutableTargetAttr targetAttr);
-
 /// Methods to get target information.
-bool isX86(IREE::HAL::ExecutableTargetAttr targetAttr);
-bool isX86_64(IREE::HAL::ExecutableTargetAttr targetAttr);
-bool isAArch64(IREE::HAL::ExecutableTargetAttr targetAttr);
-bool isRISCV(IREE::HAL::ExecutableTargetAttr targetAttr);
 bool isVMVXBackend(IREE::HAL::ExecutableTargetAttr targetAttr);
 bool hasMicrokernels(IREE::HAL::ExecutableTargetAttr targetAttr);
-bool preferIntrinsicsOverAsm(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if `targetAttr` has `feature` in its CPU features.
-bool hasFeature(IREE::HAL::ExecutableTargetAttr targetAttr, StringRef feature);
-
-/// Returns true if the 'targetAttr' contains '+avx2' in its cpu features.
-bool hasAVX2Feature(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if the 'targetAttr' contains '+v' in its cpu features.
-bool hasVFeature(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if the 'targetAttr' contains '+zve32x' in its cpu features.
-bool hasZve32xFeature(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if the 'targetAttr' contains '+zve32f' in its cpu features.
-bool hasZve32fFeature(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if the 'targetAttr' contains '+zve64x' in its cpu features.
-bool hasZve64xFeature(IREE::HAL::ExecutableTargetAttr targetAttr);
-
-/// Returns true if the 'targetAttr' contains '+sve' or '+sve2' in its cpu
-/// features.
-bool hasAnySVEFeature(IREE::HAL::ExecutableTargetAttr targetAttr);
 
 /// Checks if a tensor value is generated from a read-only object, like
 /// and interface binding with read-only attribute or from an `arith.constant`
